[X86] Add a new environment for interoperating with 32-bit code from 64-bit.
I've dubbed this new environment "wine32"--32-bit Wine on 64-bit macOS
being the primary use case for this. In this environment, 32-bit
pointers are available under address space 32. The stack is also assumed
to be in the first 4 GB of memory, to facilitate interop with 32-bit
code.

Currently, we expand these pointers to 64 bits before dereferencing
them; that's OK, but it'd be more efficient if we simply dereferenced
them as-is. Unfortunately, codegen insists on one pointer regclass to
rule them all. (Maybe `GlobalISel` will fix things...)

I'm kinda leery about the 32<->64 copies I had to implement to make this
work. I feel like we shouldn't be doing that.