[X86] Lower 64->32 downcalls to 32-bit far calls.
This lowers calls to address space 32 to a 32-bit far call from
temporary stack memory. Memory must be used because the immediate form
is not supported in 64-bit code. By default, the target segment is the
default 32-bit segment on the target OS. Later, it will become possible
to define the target segment explicitly.

Tail calls aren't supported for this type of call. The reason, of
course, is that the caller was usually called near, but the callee is
called far.