[X86] Add support for 32-bit interop calling conventions.
With one exception, these are only enabled in the 'Wine32' environment.
With these, you can call, and even implement, functions that (more or
less) conform to the 32-bit ABI.

I reused the `stdcall`, `fastcall`, and `thiscall` conventions from
x86-32, but I obviously couldn't make the `C` convention do this in
Wine32. Therefore, I added a new calling convention: `X86_64_C32`. This
new CC is available on all x86-64 targets. It has largely the same
semantics as the `C` convention on x86-32, save for the fact that 64-bit
functions use 64-bit returns (a necessity, given that there are no near
32-bit jumps allowed in 64-bit mode). On x86-32, it's exactly the same
as the `C` convention.

In addition, 64-bit functions implemented with these calling convention
do not ever pop arguments. This is to enable thunking from 32-bit to
64-bit and back again. Otherwise, the return address to the 32-bit side
would be lost. Furthermore, if called by a 64-bit function, 12 bytes of
shadow space are additionally expected to be reserved. This is where the
return address of the 32-bit caller and 32-bit thunk will be when called
from 32-bit code.

Handling 64-bit return values was a concern: since 64-bit registers are
available in 64-bit mode, LLVM wanted to use them on the assumption that
other 64-bit code would be calling this function. Unfortunately, this
required custom handling in order to stuff them in `EDX:EAX` as 32-bit
code is expecting.