[X86] Generate thunks for 64/32 interop functions.
This change adds a pass that runs late in codegen, just prior to
`AsmPrint`ing, which looks at all functions that may be used for 64/32
interop and adds thunk `MachineFunction`s for each of them.

These thunks can be used to enable a function implemented in 64-bit code
to be called by existing 32-bit code. They handle the transition to
64-bit mode and back, as well as popping the arguments if the caller is
expecting a callee-pop convention.

Thunks are *not* generated if already defined. This is true even if the
thunk is defined with module-scope inline assembly.

A concern is the way the 32-bit side pushes the target far address onto
the stack before calling it. If the `lcall` instruction pushes `%eip`
onto the stack before reading its operand, the instruction will wind up
calling the next instruction--i.e. the call will have no effect beyond
pushing 8 bytes onto the stack, and consuming about 150-300 cycles to do
a far call.