[X86] Allow both 32-bit and 64-bit pointers for memory addresses.
Only do this when compiling for the 64-32 interop target. This removes a
bunch of redundant copies that were not needed, and also fixes the
included test case, allowing it to compile successfully.

Adding the register classes to support this brings the total number of
register classes in the x86 backend, including auto-generated
subclasses, to 140. This is more than can be encoded in the single-byte
representation in the DAG matcher table. This requires TableGen to use
the VBR encoding for regclasses 128 and above, which this change does.

This also requires that the required value type be passed to the
`TargetRegisterInfo::getAllocatableClass()` method, in order to
disambiguate between 32-bit and 64-bit pointers. Otherwise, unsuitable
register classes will wind up being chosen, resulting in LLVM being
unable to allocate physical registers. I considered making the new
register classes allocatable, but that causes havoc elsewhere.